+++ pkg/nats/js_subscription_push_queue_group_handler.go
@@ -140,8 +140,11 @@ func newJsPushQueueGroupHandlerSubscription(logger *zap.Logger,
 	l := logger.Named("subscription")
 
 	var subOptions []nats.SubOpt
-	if consumerCfg.GetBackOff() != nil {
-		subOptions = append(subOptions, nats.BackOff(consumerCfg.GetBackOff()))
+	if consumerCfg.GetBackOffTimings() != nil {
+		subOptions = append(subOptions,
+			nats.BackOff(consumerCfg.GetBackOffTimings()),
+			nats.MaxDeliver(consumerCfg.GetMaxDeliveryCount()),
+		)
 	}
 
 	return &jsPushQueueGroupHandlerSubscription{